+++ community/bolt/src/test/java/org/neo4j/bolt/messaging/BoltResponseMessageReader.java
@@ -49,7 +49,7 @@ public class BoltResponseMessageReader
         try
         {
             unpacker.unpackStructHeader();
-            final int signature = (int) unpacker.unpackStructSignature();
+            final int signature = unpacker.unpackStructSignature();
             BoltResponseMessage message = BoltResponseMessage.withSignature( signature );
             try
             {
